Sync with upstream
Finally drop the distant-future-time handling proposed in pr #298, and
better handled by #302 and related updates. Note this is a *breaking
change*, and the FS revision is being pushed back to 26, to match
upstream. There is no 'upgrade'/migration process currently implemented
to go from this fork's previous approach/schema as part of this update.